requirements.txt
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
jinja2==3.1.5
sse-starlette==2.2.1
httpx==0.28.1

# Data collection
feedparser==6.0.11
beautifulsoup4==4.12.3
lxml==5.3.0

# Scheduling
apscheduler==3.11.0

# Telegram notifications
python-telegram-bot==21.10

# Database
aiosqlite==0.20.0

# Config
python-dotenv==1.0.1
pyyaml==6.0.2

# Security
pydantic==2.10.4
pydantic-settings==2.7.1
